+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Text fields can now use “title” as their API identifier if “use this field as the entry title” is ticked.
 - Field validation now returns translatable errors.
+- Crud function insert_model_entry() will now append relationship ids.
 
 ### Fixed
 - Issue where adding a new repeating field to an existing model schema could break GraphQL queries under certain conditions.